解决了现存的基本所有bug！但是效率有点低，加载界面太慢了。。
准备晚上或者明天把分数管理界面搞完，然后搞其他的

+++ src/main/java/org/example/javafx/util/CommonMethod.java
@@ -89,7 +89,7 @@ public class CommonMethod {
             return (Integer)obj;
         String str = obj.toString();
         try {
-            return (int)Double.parseDouble(str);
+            return Integer.parseInt(str);
         }catch(Exception e) {
             return null;
         }
